ssss

+++ src/freequant/FreeQuant.QuantRouterLib/Data/Level2Action.cs
@@ -1,13 +1,10 @@
-using System.Runtime.CompilerServices;
-
 namespace FreeQuant.QuantRouterLib.Data
 {
-  public class Level2Action
-  {
-    public const byte New = (byte) 0;
-    public const byte Change = (byte) 1;
-    public const byte Delete = (byte) 2;
-    public const byte Reset = (byte) 3;
-
-  }
+    public class Level2Action
+    {
+        public const byte New = 0;
+        public const byte Change = 1;
+        public const byte Delete = 2;
+        public const byte Reset = 3;
+    }
 }
