+++ src/freequant/FreeQuant.QuantRouterLib/Data/Level2Side.cs
@@ -1,11 +1,8 @@
-using System.Runtime.CompilerServices;
-
 namespace FreeQuant.QuantRouterLib.Data
 {
-  public class Level2Side
-  {
-    public const byte Bid = (byte) 0;
-    public const byte Ask = (byte) 1;
-
-  }
+    public class Level2Side
+    {
+        public const byte Bid = 0;
+        public const byte Ask = 1;
+    }
 }
